Move file definitions into their own file
